requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0

# AutoML & Meta-learning
pymfe>=0.4.0
optuna>=3.3.0

# Explainability
shap>=0.42.0
lime>=0.2.0

# LLM APIs
anthropic>=0.25.0
google-generativeai>=0.4.0

# UI
streamlit>=1.28.0

# Utils
pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
plotly>=5.17.0
